Embed lookup tables in kaitai struct
This is a workaround for kaitai-io/kaitai_struct#627. We can embed the
lookup tables for e.g. step size as computed instance values, and then
refer to these using in expressions, but this is a more cluttered
solution than we could get if tables were part of the ksy syntax.